Project structure fixed to follow steps architecture
