+++ features/environment.py
@@ -9,7 +9,7 @@ def browser_init(context):
     # context.browser = webdriver.Safari()
     # context.browser = webdriver.Firefox()
 
-    # context.driver.maximize_window()
+    context.driver.maximize_window()
     context.driver.implicitly_wait(4)
