+++ features/steps/amazon_add_to_cart.py
@@ -1,23 +1,5 @@
 from selenium.webdriver.common.by import By
-from behave import given, when, then
-from time import sleep
-
-@given('Open amazon page')
-def open_google(context):
-    context.driver.get('https://www.amazon.com/')
-
-
-@when('Input Dress into search field')
-def search_dress(context):
-    context.driver.find_element(By.ID, 'twotabsearchtextbox').send_keys('dress')
-
-
-sleep(4)
-
-
-@when('Click on search icon')
-def click_search(context):
-    context.driver.find_element(By.CSS_SELECTOR, "input.nav-input[type='submit']").click()
+from behave import when, then
 
 
 @then('Product result for dress is shown')
@@ -45,8 +27,3 @@ def click_cart_icon(context):
 def verify_item(context):
     verify_cart = context.driver.find_element(By.ID, 'sc-subtotal-label-activecart').text
     assert 'item' in verify_cart
-
-
-
-
-
