+++ features/tests/orders.feature
@@ -1,9 +1,6 @@
-
-
 Feature: Test Scenario for an ability to see Sign in page
-
 
   Scenario: User can see sign in page
     Given Open Amazon page
     When Click on orders button
-    Then Sign in page is shown
+    Then Verify Sign-in page is opened
